app/requirements.txt
############################################################
# 🧠 OCR Standalone — ONNX Only Edition (docTR + PaddleOCR)
############################################################

##############################
# ---- Base Runtime ---------
##############################
fastapi==0.115.0
uvicorn==0.30.6
python-multipart>=0.0.9
pydantic>=2.9.0
typing-extensions>=4.12.0

##############################
# ---- Image Processing -----
##############################
opencv-python-headless>=4.10.0
Pillow>=10.0.0
numpy
matplotlib>=3.9.0
torch
scipy
scikit-learn

##############################
# ---- PDF -------------------
##############################
pdf2image>=1.17.0
PyMuPDF>=1.23.8

############################################################
# 🧠 ONNX Runtime + ONNXTR (docTR-ONNX)
############################################################
onnx
onnxruntime
onnxtr @ git+https://github.com/felixdittrich92/OnnxTR.git

############################################################
# 🧠 Extraction / fuzzy matching
############################################################
rapidfuzz>=3.9.0
gliner

############################################################
# 🧠 Layout (YOLO)
############################################################
ultralytics
layoutparser

############################################################
# 🖥️ UI & Dashboard
############################################################
streamlit>=1.0.0
streamlit-autorefresh
psutil
